+++ apps/frontend/src/app/5_pages/ZeroPage/hooks/useLiquityBaseParams.ts
@@ -2,13 +2,15 @@ import { useEffect, useState } from 'react';
 
 import { Decimal } from '@sovryn/utils';
 
-import { getLiquityBaseParams } from '../utils/trove-manager';
+import { ILiquityBaseParams } from '../types';
+import { getLiquityBaseParams } from '../utils';
 
 export const useLiquityBaseParams = () => {
-  const [liquityBaseParams, setLiquityBaseParams] = useState({
-    minBorrowingFeeRate: Decimal.ZERO,
-    maxBorrowingFeeRate: Decimal.ZERO,
-  });
+  const [liquityBaseParams, setLiquityBaseParams] =
+    useState<ILiquityBaseParams>({
+      minBorrowingFeeRate: Decimal.ZERO,
+      maxBorrowingFeeRate: Decimal.ZERO,
+    });
 
   useEffect(() => {
     const getBorrowingFee = async () => {
