+++ apps/frontend/src/app/5_pages/ZeroPage/types.ts
@@ -1,0 +1,6 @@
+import { Decimal } from '@sovryn/utils';
+
+export interface ILiquityBaseParams {
+  minBorrowingFeeRate: Decimal;
+  maxBorrowingFeeRate: Decimal;
+}
